Clean diff after setup
Some of the repo_setup.sh scripts leave the working tree in a dirty state which can make it difficult to generate a patch that applies cleanly.  This change commits any outstanding changges such that any patch generated with `git diff` will cleanly apply to a newly launched container during the evaluation step.